requirements.txt
torch
torchvision
torchaudio
xformers
bitsandbytes
accelerate
peft
trl
transformers>=4.40.0

unsloth[colab-new] @ git+https://github.com/unslothai/unsloth.git

llamafactory-cli

f5-tts
soundfile
librosa
jiwer

datasets
pandas
scipy
python-dotenv
pyyaml
wandb
tqdm